hf-space/requirements.txt
aiohttp==3.9.1
cffi==1.16.0
beautifulsoup4==4.12.2
tls-client==1.0.1
colorama==0.4.6
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv
schedule==1.2.1
openai
groq
pydantic>=2.0
python-jose[cryptography]
passlib[bcrypt]
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
databases
pydantic>=2.0
pydantic-settings
email-validator
python-multipart
python-dotenv
aiohttp
requests
pandas
APScheduler
numpy
beautifulsoup4
lxml
openpyxl
yfinance>=0.2.50
yahooquery>=2.3.0
cachetools
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
pyyaml>=6.0
httpx>=0.25.0
playwright>=1.40.0
aiosmtplib
curl-cffi==0.5.10
